+++ bibfor/algeline/ckatrc.F90
@@ -24,7 +24,7 @@ subroutine ckatrc(vr, xsi, coefck, ck)
 #include "asterfort/routhc.h"
     real(kind=8) :: vr, xsi, coefck(1, 11), ck
 !
-    integer :: ior
+    integer(kind=8) :: ior
     real(kind=8) :: dcldy, a0(2), dr(2)
     real(kind=8) :: pi, hr, hi, xcor, omr
     complex(kind=8) :: pr
