+++ bibfor/algeline/colni1.F90
@@ -20,10 +20,10 @@ subroutine colni1(col, n, d1, t, eps, &
                   ier)
 ! person_in_charge: olivier.boiteau at edf.fr
     implicit none
-    integer :: n, ier
+    integer(kind=8) :: n, ier
     real(kind=8) :: col(n), d1, t(n), eps
 !
-    integer :: i
+    integer(kind=8) :: i
 !-----------------------------------------------------------------------
 !-----------------------------------------------------------------------
     if (abs(d1) .le. eps) then
