+++ bibfor/algeline/crsint.F90
@@ -38,9 +38,9 @@ subroutine crsint(solveu)
 !-----------------------------------------------------------------------
 !     VARIABLES LOCALES
 !----------------------------------------------------------------------
-    integer :: zslvk, zslvr, zslvi
-    integer :: islvk, islvr, islvi
-    integer :: i, monit(12), niv, nbproc, rang, ifm, iret
+    integer(kind=8) :: zslvk, zslvr, zslvi
+    integer(kind=8) :: islvk, islvr, islvi
+    integer(kind=8) :: i, monit(12), niv, nbproc, rang, ifm, iret
     mpi_int :: mrank, msize
     character(len=24) :: kmonit(12)
 !----------------------------------------------------------------------
